Update with PSU information
Add PSU experience and education. Clean up some of the existing lines,
like removing the EnSURE and continued research lines from the two REUs.
Changed the 'programming languages' section to be computational tools.
Condensed the mentoring information in the resume down to just five lines
that show the classes and two of those lines are to show the students and
responsibilities. Updated to PSU email.